Adding versions compatible for font awesome. The 3 ready versions or just bootstrap missing the icon css, and the included file has most of the css for font awesome included. It is missing the ie7 css support and is missing the actual @font-face call:
@font-face {
  font-family: "FontAwesome";
  src: url('../font/fontawesome-webfont.eot');
  src: url('../font/fontawesome-webfont.eot?#iefix') format('eot'), url('../font/fontawesome-webfont.woff') format('woff'), url('../font/fontawesome-webfont.ttf') format('truetype'), url('../font/fontawesome-webfont.svg#FontAwesome') format('svg');
  font-weight: normal;
  font-style: normal;
}

I left this out has users may want to have the font installed in a different directory. If we add the font to the cdn maybe we can re-add this code with the path to the cdn url. Not sure where you would want to place the fonts in your repo. But we can grab the font files via a git submodule from the repo: https://github.com/FortAwesome/Font-Awesome