appveyor.yml: enable symlinks for git.
This fixes the different number of tests on AppVeyor. It's because in the repository we have a `bootstrap` symlink, and by default git symlink support is disabled on Windows.

Note that one needs to clone the repository with admin rights for the symlinks to work. AppVeyor does this by default.